+++ src/Pim/Bundle/CatalogBundle/PimCatalogBundle.php
@@ -20,24 +20,6 @@ class PimCatalogBundle extends Bundle
     const DOCTRINE_MONGODB = '\Doctrine\Bundle\MongoDBBundle\DependencyInjection\Compiler\DoctrineMongoDBMappingsPass';
 
     /** @staticvar string */
-    const VERSION = '1.2.0-DEV';
-
-    /** @staticvar string */
-    const VERSION_CODENAME = '';
-
-    /** @staticvar string */
-    const MAJOR_VERSION = '1';
-
-    /** @staticvar string */
-    const MINOR_VERSION = '2';
-
-    /** @staticvar string */
-    const PATCH_VERSION = '0';
-
-    /** @staticvar string */
-    const EXTRA_VERSION = '';
-
-    /** @staticvar string */
     const ODM_ENTITIES_TYPE = 'entities';
 
     /** @staticvar string */